Initial commit
This is the initial commit of Martin Bengtsson's code which is the base
of this project.  All credit is given to Martin Bengtsson for the
oringal code which uses the MIT license.  The orignal code can be found
at https://gallery.technet.microsoft.com/Windows-10-Toast-9f228eb1

WIP:  added variables to capture Toast Actions

WIP: refactored out opening Toast tag in order to build it up

WIP: Toast tag provided attributes for Toast Action

WIP: no sense have a Toast Action if it is not enabled